calims2-webapp/src/java/gov/nih/nci/calims2/ui/administration/customerservice/contract/ContractController.java
/*L
 * Copyright Moxie Informatics.
 *
 * Distributed under the OSI-approved BSD 3-Clause License.
 * See http://ncip.github.com/calims/LICENSE.txt for details.
 */

/**
 *
 */
package gov.nih.nci.calims2.ui.administration.customerservice.contract;

import java.util.Locale;

import javax.annotation.Resource;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.ModelAttribute;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.servlet.ModelAndView;

import gov.nih.nci.calims2.business.generic.GenericService;
import gov.nih.nci.calims2.domain.administration.customerservice.Contract;
import gov.nih.nci.calims2.domain.administration.customerservice.enumeration.ContractStatus;
import gov.nih.nci.calims2.ui.generic.crud.CRUDController;
import gov.nih.nci.calims2.ui.generic.crud.CRUDControllerConfig;
import gov.nih.nci.calims2.ui.generic.crud.CRUDFormDecorator;
import gov.nih.nci.calims2.uic.descriptor.table.Table;
import gov.nih.nci.calims2.util.enumeration.I18nEnumeration.I18nEnumerationHelper;

/**
 * @author katariap
 * 
 */
@Controller
@RequestMapping(ContractController.URL_PREFIX)
public class ContractController extends CRUDController<Contract> {

  /** Prefix of urls of this module. */
  static final String URL_PREFIX = "/administration/customerservice/contract";
  
  /**
   * Default constructor.
   */
  public ContractController() {
    super(URL_PREFIX, "name");
    CRUDControllerConfig config = getConfig();
    config.setSubFlowUrls(new String[] {"/common/document/enterList.do"});
    config.setAdvancedSearch(true);
  }
  
  /**
   * {@inheritDoc}
   */
  public ModelAndView completeEditModel(ModelAndView model, Locale locale) {
    model.addObject("contractStatuses", I18nEnumerationHelper.getLocalizedValues(ContractStatus.class, locale));
    return model;
  }

  /**
   * Saves an entity and returns a refreshed list type page.
   * 
   * @param form The object containing the values of the entity to be saved.
   * @param locale The current locale.
   * @return The list view.
   */
  @RequestMapping(CRUDControllerConfig.STD_SAVE_COMMAND)
  public ModelAndView save(@ModelAttribute("form") ContractForm form, Locale locale) {
    return doSave(form, locale);
  }

  /**
   * @param contractService the contractService to set
   */
  @Resource(name = "contractService")
  public void setContractService(GenericService<Contract> contractService) {
    super.setMainService(contractService);
  }
  
  /**
   * @param formDecorator the formDecorator to set
   */
  @Resource(name = "contractFormDecorator")
  public void setFormDecorator(CRUDFormDecorator formDecorator) {
    super.setFormDecorator(formDecorator);
  }
  
  /**
   * @param listTable the listTable to set
   */
  @Resource(name = "contractTable")
  public void setListTable(Table listTable) {
    super.setListTable(listTable);
  }
}
